Fix 'on' parameter in pp.merge_with_ir.
Previously, when 'on' was set to None (the default) it would have passed
to pandas:
 - "batch", when the column is present in both `obs` data frames, and
 - None, otherwise.

The intention was that in the latter case the join is only performed on
the index. However, pandas interprets None as the "intersection of all
columns present in both dataframes".

This commit changes this explicitly to only join on the index.