Update merge_coo_matrices empty list special case
Previously, the function returned 0 to match the behavior
of `sum`. Now it returns an empty coo matrix.

Possibly I need to change the behaviour to return
a row of zeros instead in the future. This would be
the most consistent behaviour but requires passing
the shape of the list additionally, that's why I would
like to avoid it.