Simplify folder structure and update API docs
This PR renames _preprocessing to pp, _tools to tl etc.

This is a relict from the early scanpy API, but now
not even scanpy uses the hidden modules anymore and
I think directly using modules is the cleanest option
and solves an issue with sphinx (couldn't handle function aliases)
and made a few other problems previously.