Change clonotype calling behavior for missing cdr3 sequences
Previously, cells that had a receptor, but no sequence
were treated differently from cells with no receptor: Previously
cells with a receptor, but no sequence were assigned to a separate
clonotype, while cells without a receptor got the clonotype `nan`.

Now, also cells without sequence are assigned the clonotype `nan`.

In practice, this shouldn't have affected a lot of people, as
during IO, it was anyway ensured that only chains with a sequence
are imported.