Fix reverse key lookup
In the "match_gene_segment" function in _clonotype_neighbors, a "keys
array" is created to map from indicies in the reverse lookup table
(these refer to rows in the clonotype table) to numeric indices of
the reverse lookup table.

The size of this keys array was defined as the size of the clonotype
table. Instead it should have been the size of the distance matrix.

For real-world cases, this hardly was an issue, since the number
of clonotypes is typically larger than the number of columns
in the distance matrix.